feat(websocket) add websocket game
