+++ src/main/java/com/souma1024/shogiv2/websocket/dto/ReconnectResponse.java
@@ -1,4 +1,4 @@
-package com.souma1024.shogiv2.websocket;
+package com.souma1024.shogiv2.websocket.dto;
 
 import com.souma1024.shogiv2.websocket.dto.enums.WebSocketType;
